add free to one philo

+++ philo/philo/func_utils.c
@@ -6,7 +6,7 @@
 /*   By: amoukhle <amoukhle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/15 12:37:22 by amoukhle          #+#    #+#             */
-/*   Updated: 2023/05/17 12:23:47 by amoukhle         ###   ########.fr       */
+/*   Updated: 2023/05/29 20:37:15 by amoukhle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,10 +50,10 @@ int	ft_isdigit(int c)
 		return (0);
 }
 
-int	ft_atoi(char *str)
+unsigned long	ft_atoi(char *str)
 {
-	int					c;
-	unsigned long long	res;
+	int				c;
+	unsigned long	res;
 
 	c = 0;
 	res = 0;
